// src/app/models/pdpa-privacy/pdpa-privacy-create/pdpa-privacy-create-request.ts
export interface PdpaPrivacyCreateRequest {
    ppCode: string;
    status: number;
    version: string;
    titleTh: string;
    titleEn: string;
    titleZh: string;
    descriptionTh: string;
    descriptionEn: string;
    descriptionZh: string;
    creatorUserId: number;
    isDeleted: number;
}